Revamped interpretation of OSM data primitives. Previously:  - each node on a way was a routing node  - each segment of a way was a routing edge Now:  - nodes adjacent to more than one way are routing nodes  - the first and last node of a way are routing nodes  - a routing a is a path of way segments between two routing nodes
This means the IRouter plugin has to deal with less nodes and edges, but
with more complicated edge descriptions. This also paves the way for
edge properties to be returned by the IRouter plugin in the future.

API changes:
IImporter: Added support for edges' description. Added support for the
IRouter plugin to remap edge IDs. An edge is now uniquely identified by
( source, target, id ). Added preliminary support for way names and
type. There is no way to retrieve them from the IRouter plugin, but I
will work on that next.
IRenderer: Renamed a member function
IGPSLookup: Returns only the nearest way, but also the way's description
as well as the number of way coordinates before the nearest point.

OMSImporter: Major overhaul
CH + CHC: Preprocessing takes up less memory due to less routing nodes.
The graph takes up less storage but the shortcuts' and edges'
descriptions take up more space. The as ( source, target ) identifies an
edge most of the time the edgeIDs are mostly 0 or 1;
GPSGrid: Stores the edges' description in addition to the edges
OSMRenderer: Minor changes to account for minor IRouter changes
Client + Preprocessor + RoutingDaemon: Minor changes

This revision will most likely introduce several minor and / or major
bugs. Please feel free to post an issue. If anyone is interested in
becoming a regular beta tester please contact me.

As I will a way names and type to the IRouter's result in the next weeks
expect some more API changes.