IRouter: Added seconds to the Edge description to allow for more precise placement / hiding <-> showing of routing instructions
CH: Encoded seconds with an error of 1/32 + 0.5s. Only 8 additional bits
are needed and were already allocated in the unpacked shortcuts path
desciption -> storage requirements stay the same

utils: added functions to encode / decode an integer in a floating point
like manner. E.g., 0 - 2^20 can be encoded in 8 bits ( 255 distinct
states ) with an error of 1/32