Client:
Added map package selection via clicking on the bounding box on a world
map. Uses raster images as a backdrop for now. They will have to be
adjusted to the memory limitations of the target device ( makes no sense
unpacking a 2048x1024 bitmap on Windows Mobile ). Maybe a vector
representation might be better, but I am not sure about that.

The displayed area is zoomed in on the bounding box of all available
data sets. Works very well when data sets are not to far away, but might
not be so good when having a data set of the US and China at the same
time.

Of course, it is still possible to select a map package from a list of
all available packages.